+++ pkg/clientconn/client.go
@@ -1,13 +1,14 @@
 package clientconn
 
 import (
+	"context"
 	"crypto/tls"
 	"crypto/x509"
 	"time"
 
 	"github.com/pkg/errors"
 	"github.com/stackrox/rox/pkg/errorhelpers"
-	pkgGRPC "github.com/stackrox/rox/pkg/grpc"
+	"github.com/stackrox/rox/pkg/grpc/alpn"
 	"github.com/stackrox/rox/pkg/grpc/authn/basic"
 	"github.com/stackrox/rox/pkg/grpc/authn/servicecerttoken"
 	"github.com/stackrox/rox/pkg/grpc/authn/tokenbased"
@@ -20,13 +21,44 @@ import (
 )
 
 var (
-	nextProtos = []string{pkgGRPC.PureGRPCALPNString, "h2"}
+	// NextProtos are the ALPN protos to use for gRPC connections.
+	NextProtos = []string{alpn.PureGRPCALPNString, "h2"}
 )
+
+// Options specifies options for establishing a gRPC client connection.
+type Options struct {
+	TLS         TLSConfigOptions
+	PerRPCCreds credentials.PerRPCCredentials
+
+	DialTLS DialTLSFunc
+}
+
+func (o *Options) dialTLSFunc() DialTLSFunc {
+	if o.DialTLS != nil {
+		return o.DialTLS
+	}
+	return DialTLS
+}
+
+func (o *Options) tlsConfig(server mtls.Subject) (*tls.Config, error) {
+	return TLSConfig(server, o.TLS)
+}
+
+// ConfigureBasicAuth configures this client connection to authenticate via basic authentication.
+func (o *Options) ConfigureBasicAuth(username, password string) {
+	o.PerRPCCreds = basic.PerRPCCredentials(username, password)
+}
+
+// ConfigureTokenAuth configures this client connection to authenticate via token-based authentication.
+func (o *Options) ConfigureTokenAuth(token string) {
+	o.PerRPCCreds = tokenbased.PerRPCCredentials(token)
+}
 
 // TLSConfigOptions are options that modify the behavior of `TLSConfig`.
 type TLSConfigOptions struct {
-	UseClientCert bool
-	ServerName    string
+	UseClientCert      bool
+	ServerName         string
+	InsecureSkipVerify bool
 }
 
 // verifyPeerCertificateFunc returns a function that can be used as the `VerifyPeerCertificate` callback of a
@@ -85,11 +117,6 @@ func verifyPeerCertificateFunc(server mtls.Subject, serviceCA *x509.CertPool, se
 
 // TLSConfig returns a TLS config that can be used to talk to the given server via mTLS.
 func TLSConfig(server mtls.Subject, opts TLSConfigOptions) (*tls.Config, error) {
-	serviceCA, err := verifier.TrustedCertPool()
-	if err != nil {
-		return nil, errors.Wrap(err, "trusted pool")
-	}
-
 	serverName := opts.ServerName
 	if serverName == "" {
 		serverName = server.Hostname()
@@ -97,8 +124,7 @@ func TLSConfig(server mtls.Subject, opts TLSConfigOptions) (*tls.Config, error) 
 
 	conf := &tls.Config{
 		ServerName: serverName,
-		RootCAs:    serviceCA,
-		NextProtos: nextProtos,
+		NextProtos: NextProtos,
 	}
 
 	if opts.UseClientCert {
@@ -109,16 +135,26 @@ func TLSConfig(server mtls.Subject, opts TLSConfigOptions) (*tls.Config, error) 
 		conf.Certificates = []tls.Certificate{clientCert}
 	}
 
-	if serverName != server.Hostname() {
-		// Since we want to support verifying against both a StackRox Service CA or a System CA, we don't know the
-		// ServerName against which to verify. While we could leave the `ServerName` field empty and only check the
-		// `ServerName` of the verified chains passed to the `VerifyPeerCertificate` callback, this would have the
-		// undesired side effect of not sending SNI in the handshake. Hence, skip the verification done by the tls
-		// library altogether and do everything in our `VerifyPeerCertificate` callback.
-		// Don't worry - this looks scary, but is actually not insecure; just slightly more flexible than what the
-		// tls library supports natively.
+	if !opts.InsecureSkipVerify {
+		serviceCA, err := verifier.TrustedCertPool()
+		if err != nil {
+			return nil, errors.Wrap(err, "trusted pool")
+		}
+		conf.RootCAs = serviceCA
+
+		if serverName != server.Hostname() {
+			// Since we want to support verifying against both a StackRox Service CA or a System CA, we don't know the
+			// ServerName against which to verify. While we could leave the `ServerName` field empty and only check the
+			// `ServerName` of the verified chains passed to the `VerifyPeerCertificate` callback, this would have the
+			// undesired side effect of not sending SNI in the handshake. Hence, skip the verification done by the tls
+			// library altogether and do everything in our `VerifyPeerCertificate` callback.
+			// Don't worry - this looks scary, but is actually not insecure; just slightly more flexible than what the
+			// tls library supports natively.
+			conf.InsecureSkipVerify = true
+			conf.VerifyPeerCertificate = verifyPeerCertificateFunc(server, serviceCA, serverName)
+		}
+	} else {
 		conf.InsecureSkipVerify = true
-		conf.VerifyPeerCertificate = verifyPeerCertificateFunc(server, serviceCA, serverName)
 	}
 
 	// If the ServerName is an IP address, no SNI will be sent by the client. Because some SNI is always better than no
@@ -133,6 +169,7 @@ func TLSConfig(server mtls.Subject, opts TLSConfigOptions) (*tls.Config, error) 
 
 type connectionOptions struct {
 	useServiceCertToken bool
+	dialTLSFunc         DialTLSFunc
 }
 
 // ConnectionOption allows specifying additional options when establishing GRPC connections.
@@ -154,6 +191,14 @@ func UseServiceCertToken(use bool) ConnectionOption {
 	})
 }
 
+// UseDialTLSFunc uses the given connection function for dialing.
+func UseDialTLSFunc(fn DialTLSFunc) ConnectionOption {
+	return connectOptFunc(func(opts *connectionOptions) error {
+		opts.dialTLSFunc = fn
+		return nil
+	})
+}
+
 // AuthenticatedGRPCConnection returns a grpc.ClientConn object that uses
 // client certificates found on the local file system.
 func AuthenticatedGRPCConnection(endpoint string, server mtls.Subject, extraConnOpts ...ConnectionOption) (conn *grpc.ClientConn, err error) {
@@ -169,18 +214,12 @@ func AuthenticatedGRPCConnection(endpoint string, server mtls.Subject, extraConn
 		return nil, errors.Wrapf(err, "could not parse endpoint %q", endpoint)
 	}
 
-	tlsConfig, err := TLSConfig(server, TLSConfigOptions{
-		UseClientCert: true,
-		ServerName:    host,
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	creds := credentials.NewTLS(tlsConfig)
-	dialOpts := []grpc.DialOption{
-		grpc.WithTransportCredentials(creds),
-		keepAliveDialOption(),
+	clientConnOpts := Options{
+		TLS: TLSConfigOptions{
+			UseClientCert: true,
+			ServerName:    host,
+		},
+		DialTLS: connOpts.dialTLSFunc,
 	}
 
 	if connOpts.useServiceCertToken {
@@ -188,31 +227,25 @@ func AuthenticatedGRPCConnection(endpoint string, server mtls.Subject, extraConn
 		if err != nil {
 			return nil, errors.Wrap(err, "loading client certificate")
 		}
-		dialOpts = append(dialOpts, grpc.WithPerRPCCredentials(servicecerttoken.NewServiceCertClientCreds(&leafCert)))
+		clientConnOpts.PerRPCCreds = servicecerttoken.NewServiceCertClientCreds(&leafCert)
 	}
 
-	return grpc.Dial(endpoint, dialOpts...)
+	return GRPCConnection(context.Background(), server, endpoint, clientConnOpts, keepAliveDialOption())
 }
 
-// GRPCConnectionWithBasicAuth returns a grpc.ClientConn using the given username/password to authenticate
-// via basic auth.
-func GRPCConnectionWithBasicAuth(endpoint string, serverName, username, password string) (*grpc.ClientConn, error) {
-	return grpcConnectionWithPerRPCCreds(endpoint, serverName, basic.PerRPCCredentials(username, password))
-}
-
-// GRPCConnectionWithToken returns a grpc.ClientConn using the given token to authenticate
-func GRPCConnectionWithToken(endpoint, serverName, token string) (*grpc.ClientConn, error) {
-	return grpcConnectionWithPerRPCCreds(endpoint, serverName, tokenbased.PerRPCCredentials(token))
-}
-
-func grpcConnectionWithPerRPCCreds(endpoint string, serverName string, perRPCCreds credentials.PerRPCCredentials) (*grpc.ClientConn, error) {
-	tlsConfig := &tls.Config{
-		InsecureSkipVerify: true,
-		ServerName:         serverName,
-		NextProtos:         nextProtos,
+// GRPCConnection establishes a gRPC connection to the given server, using the given connection options.
+func GRPCConnection(dialCtx context.Context, server mtls.Subject, endpoint string, clientConnOpts Options, dialOpts ...grpc.DialOption) (*grpc.ClientConn, error) {
+	tlsConf, err := clientConnOpts.tlsConfig(server)
+	if err != nil {
+		return nil, errors.Wrap(err, "instantiating TLS config")
 	}
-	creds := credentials.NewTLS(tlsConfig)
-	return grpc.Dial(endpoint, grpc.WithTransportCredentials(creds), grpc.WithPerRPCCredentials(perRPCCreds))
+
+	allDialOpts := make([]grpc.DialOption, 0, len(dialOpts)+1)
+	if clientConnOpts.PerRPCCreds != nil {
+		allDialOpts = append(allDialOpts, grpc.WithPerRPCCredentials(clientConnOpts.PerRPCCreds))
+	}
+	allDialOpts = append(allDialOpts, dialOpts...)
+	return clientConnOpts.dialTLSFunc()(dialCtx, endpoint, tlsConf, allDialOpts...)
 }
 
 // Parameters for keep alive.