+++ pkg/grpc/alpn/alpn.go
@@ -1,4 +1,4 @@
-package grpc
+package alpn
 
 import (
 	"crypto/tls"