+++ pkg/grpcweb/defs.go
@@ -1,0 +1,8 @@
+package grpcweb
+
+const (
+	compressedFlag     byte = 1 << 0
+	trailerMessageFlag byte = 1 << 7
+
+	completeHeaderLen = 5
+)