+++ roxctl/common/flags/endpoint.go
@@ -7,6 +7,7 @@ import (
 var (
 	endpoint   string
 	serverName string
+	directGRPC bool
 )
 
 // AddEndpoint adds the endpoint flag to the base command.
@@ -19,6 +20,11 @@ func AddServerName(c *cobra.Command) {
 	c.PersistentFlags().StringVarP(&serverName, "server-name", "s", "", "TLS ServerName to use for SNI (if empty, derived from endpoint)")
 }
 
+// AddDirectGRPC adds the direct-grpc flag to the command.
+func AddDirectGRPC(c *cobra.Command) {
+	c.PersistentFlags().BoolVar(&directGRPC, "direct-grpc", false, "Use direct gRPC (advanced; only use if you encounter connection issues)")
+}
+
 // Endpoint returns the set endpoint.
 func Endpoint() string {
 	return endpoint
@@ -27,4 +33,9 @@ func Endpoint() string {
 // ServerName returns the specified ServerName.
 func ServerName() string {
 	return serverName
+}
+
+// UseDirectGRPC returns whether to use gRPC directly, i.e., without a proxy.
+func UseDirectGRPC() bool {
+	return directGRPC
 }