+++ roxctl/main.go
@@ -58,6 +58,7 @@ func main() {
 	flags.AddPassword(c)
 	flags.AddEndpoint(c)
 	flags.AddServerName(c)
+	flags.AddDirectGRPC(c)
 
 	if err := c.Execute(); err != nil {
 		os.Exit(1)