+++ scripts/ci/proxy/deploy.sh
@@ -15,18 +15,25 @@ kubectl -n proxies create cm nginx-proxy-plain-http-conf --from-file="${dir}/ngi
 
 kubectl apply -f "${dir}/nginx-proxy-plain-http.yaml"
 
-kubectl -n proxies create cm nginx-proxy-tls-multiplexed-conf \
-	--from-file="${dir}/nginx-proxy-tls-multiplexed.conf" \
-	--dry-run -o yaml | kubectl apply -f -
-
 cert_dir="${PROXY_CERTS_DIR:-$(mktemp -d)}"
 "${dir}/../../../tests/scripts/setup-certs.sh" "${cert_dir}" "central-proxy.stackrox.local" "Proxy CA"
 
-kubectl -n proxies create secret tls nginx-proxy-tls-multiplexed-certs \
+kubectl -n proxies create secret tls nginx-proxy-tls-certs \
 	--cert="${cert_dir}/tls.crt" \
 	--key="${cert_dir}/tls.key" \
 	--dry-run -o yaml | kubectl apply -f -
 
-kubectl apply -f "${dir}/nginx-proxy-tls-multiplexed.yaml"
+for proxy_type in http1 http1-plain http2 http2-plain multiplexed multiplexed-tls-be; do
+	proxy_name="nginx-proxy-tls-${proxy_type}"
+
+	kubectl -n proxies create cm "${proxy_name}-conf" \
+		--from-file="${dir}/${proxy_name}.conf" \
+		--dry-run -o yaml | kubectl apply -f -
+
+	kubectl apply -n proxies -f <(name="${proxy_name}" envsubst <"${dir}/nginx-proxy-tls.yaml.template")
+done
+
 sleep 5
-kubectl -n proxies wait --for=condition=available deploy/nginx-proxy-{plain-http,tls-multiplexed} --timeout=2m
+kubectl -n proxies wait --for=condition=available \
+	deploy/nginx-proxy-{plain-http,tls-multiplexed,tls-http1,tls-http1-plain,tls-http2,tls-http2-plain,tls-multiplexed-tls-be} \
+	--timeout=2m